+++ tests/covalent_tests/results_manager_tests/write_result_to_db_test.py
@@ -1,0 +1,61 @@
+# Copyright 2021 Agnostiq Inc.
+#
+# This file is part of Covalent.
+#
+# Licensed under the GNU Affero General Public License 3.0 (the "License").
+# A copy of the License may be obtained with this software package or at
+#
+#      https://www.gnu.org/licenses/agpl-3.0.en.html
+#
+# Use of this file is prohibited except in compliance with the License. Any
+# modifications or derivative works of this file must retain this copyright
+# notice, and modified files must contain a notice indicating that they have
+# been altered from the originals.
+#
+# Covalent is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the License for more details.
+#
+# Relief from the License may be granted by purchasing a commercial license.
+
+"""Unit tests for the module used to write the decomposed result object to the database."""
+
+
+def test_insert_lattices_data():
+    pass
+
+
+def test_insert_electrons_data():
+    pass
+
+
+def test_insert_electron_dependency_data():
+    pass
+
+
+def test_update_lattices_data():
+    pass
+
+
+def test_update_electrons_data():
+    pass
+
+
+def test_get_electron_dependencies():
+    pass
+
+
+def test_sublattice_electron_id():
+    pass
+
+
+def test_are_electron_dependencies_added():
+    pass
+
+
+def test_is_lattice_added():
+    pass
+
+
+def test_get_electron_type():
+    pass